battery handling / fixed wakeup pref sync
Added default battery functions.
Fixed issue where wakeup wasn't actually syncing preferences.   How did that ever work?
