+++ Consolidated Drivers/zwave-sensor/src/init.lua
@@ -61,6 +61,8 @@ end
 --- @param self st.zwave.Driver
 --- @param device st.zwave.Device
 local function device_init(self, device)
+  -- Set an update preferences function.  The framework will call this
+  -- in the default wakeup handlers if configuration is needed to be delayed set.
   device:set_update_preferences_fn(preferences.update_preferences)
 end
 
@@ -154,6 +156,7 @@ local driver_template = {
   },
   sub_drivers = {
     require("sleepy-device"),  -- General support for any sleepy zwave devices
+    require("battery-device"), -- Support for zwave battery devices
     require("ecolink-sensor"),
     require("fortrezz-leak"),
     require("homeseer-leak"),
